Make ApiCall function to set unread notifi to read
