+++ client/src/App.tsx
@@ -13,6 +13,7 @@ import Booking from './pages/Booking/Booking';
 import { AuthProvider } from './context/useAuthContext';
 import { SocketProvider } from './context/useSocketContext';
 import { SnackBarProvider } from './context/useSnackbarContext';
+import { NotificationProvider } from './context/useNotificationContext';
 
 import './App.css';
 
@@ -23,22 +24,24 @@ function App(): JSX.Element {
         <SnackBarProvider>
           <AuthProvider>
             <SocketProvider>
-              <NavBar />
-              <Switch>
-                <Route exact path="/login" component={Login} />
-                <Route exact path="/signup" component={Signup} />
-                <Route path="/profile" component={Profile} />
-                <Route exact path="/my-jobs" component={MyJobs} />
-                <Route exact path="/my-sitters" component={MySitters} />
-                <Route exact path="/messages" component={Messages} />
-                <Route exact path="/dashboard">
-                  <Dashboard />
-                </Route>
-                <Route path="/booking" component={Booking} />
-                <Route path="*">
-                  <Redirect to="/login" />
-                </Route>
-              </Switch>
+              <NotificationProvider>
+                <NavBar />
+                <Switch>
+                  <Route exact path="/login" component={Login} />
+                  <Route exact path="/signup" component={Signup} />
+                  <Route path="/profile" component={Profile} />
+                  <Route exact path="/my-jobs" component={MyJobs} />
+                  <Route exact path="/my-sitters" component={MySitters} />
+                  <Route exact path="/messages" component={Messages} />
+                  <Route exact path="/dashboard">
+                    <Dashboard />
+                  </Route>
+                  <Route path="/booking" component={Booking} />
+                  <Route path="*">
+                    <Redirect to="/login" />
+                  </Route>
+                </Switch>
+              </NotificationProvider>
             </SocketProvider>
           </AuthProvider>
         </SnackBarProvider>
