+++ client/src/helpers/APICalls/notifications.ts
@@ -1,0 +1,18 @@
+// import { NotificationApiData } from '../../interface/NotificationApiData';
+
+const getUnreadNotifications = async () => {
+  return await fetch(`/notification/unread`)
+    .then((res) => res.json())
+    .catch(() => ({
+      error: { message: 'Unable to connect to server. Please try again' },
+    }));
+};
+const setUnreadNotificationToRead = async (id: string) => {
+  return await fetch(`/notification/${id}`, { method: 'PATCH' })
+    .then((res) => res.json())
+    .catch(() => ({
+      error: { message: 'Unable to connect to server. Please try again' },
+    }));
+};
+
+export { getUnreadNotifications, setUnreadNotificationToRead };
